Added a source file for `JoltTempAllocator`
